supports the enum attrs, test=develop
